Refactor Helpers into separate files
