+++ Documentation/Demo/Dithering/Extensions/AnyPublisher+Extension.swift
@@ -1,0 +1,16 @@
+//
+//  AnyPublisher+Extension.swift
+//  Dithering
+//
+//  Created by Eskil Gjerde Sviggum on 11/03/2025.
+//
+
+import Combine
+
+extension AnyPublisher {
+
+    static func empty() -> AnyPublisher<Any, Never> {
+        CurrentValueSubject<Any, Never>(0).eraseToAnyPublisher()
+    }
+    
+}
