+++ Documentation/Demo/Dithering/Extensions/Numeric+Extension.swift
@@ -1,0 +1,19 @@
+//
+//  Numeric+Extension.swift
+//  Dithering
+//
+//  Created by Eskil Gjerde Sviggum on 11/03/2025.
+//
+
+
+extension Numeric {
+    
+    func add1IfZero() -> Self {
+        if self == .zero {
+            return 1
+        }
+        
+        return self
+    }
+    
+}
