+++ Documentation/Demo/Dithering/Extensions/UIColor+Extension.swift
@@ -1,0 +1,16 @@
+//
+//  UIColor+Extension.swift
+//  Dithering
+//
+//  Created by Eskil Gjerde Sviggum on 11/03/2025.
+//
+
+import SwiftUI
+
+extension UIColor {
+    
+    func toColor() -> Color {
+        Color(uiColor: self)
+    }
+    
+}
