+++ Documentation/Demo/Dithering/Helpers/SafeAreaInsets+EnvironmentValues.swift
@@ -1,0 +1,24 @@
+//
+//  SafeAreaInsets+EnvironmentValues.swift
+//  Dithering
+//
+//  Created by Eskil Gjerde Sviggum on 11/03/2025.
+//
+
+import SwiftUI
+
+private struct SafeAreaInsetsKey: EnvironmentKey {
+    static var defaultValue: EdgeInsets {
+        let scene = UIApplication.shared.connectedScenes.first as? UIWindowScene
+        let window = scene?.windows.first
+        let uiInsets = window?.safeAreaInsets ?? .zero
+        return EdgeInsets(top: uiInsets.top, leading: uiInsets.left, bottom: uiInsets.bottom, trailing: uiInsets.right)
+    }
+}
+
+extension EnvironmentValues {
+    
+    var safeAreaInsets: EdgeInsets {
+        self[SafeAreaInsetsKey.self]
+    }
+}
